+++ python/tests/integration/sdk/test_cluster_ops.py
@@ -8,7 +8,6 @@ import unittest
 
 from aistore.sdk import Client
 from aistore.sdk.const import ACT_COPY_OBJECTS
-from aistore.sdk.types import ClusterPerformance
 from tests.integration import CLUSTER_ENDPOINT
 from tests.utils import random_string
 
@@ -88,15 +87,8 @@ class TestClusterOps(unittest.TestCase):  # pylint: disable=unused-variable
         smap = self.cluster.get_info()
         performance = self.cluster.get_performance()
 
-        self.assertIsInstance(performance, ClusterPerformance)
-
-        num_targets_in_smap = len(smap.tmap)
-
-        self.assertEqual(num_targets_in_smap, len(performance.throughput))
-        self.assertEqual(num_targets_in_smap, len(performance.latency))
-        self.assertEqual(num_targets_in_smap, len(performance.counters))
-
+        self.assertIsInstance(performance, dict)
+        self.assertEqual(len(smap.tmap), len(performance))
         for target_id in smap.tmap:
-            self.assertIn(target_id, performance.throughput)
-            self.assertIn(target_id, performance.latency)
-            self.assertIn(target_id, performance.counters)
+            self.assertIn(target_id, performance)
+            self.assertIsInstance(performance[target_id], dict)